moved easily moved API handlers out of main
