+++ pkg/web/handlers_trackinfo.go
@@ -1,0 +1,49 @@
+package web
+
+import (
+	"encoding/json"
+	"net/http"
+	"strings"
+)
+
+// trackInfo returns lightweight upload metadata for a track. GET /api/track-info/{trackID}
+func (s *Server) trackInfo(w http.ResponseWriter, r *http.Request) {
+	if s.DB == nil || s.DB.DB == nil {
+		http.Error(w, "Database not available", http.StatusServiceUnavailable)
+		return
+	}
+	trackID := strings.TrimPrefix(r.URL.Path, "/api/track-info/")
+	if trackID == "" {
+		http.Error(w, "Missing track ID", http.StatusBadRequest)
+		return
+	}
+	ctx := r.Context()
+	var username, detector string
+	var recordingDate int64
+	var query string
+	switch s.Config.DBType {
+	case "pgx", "duckdb":
+		query = `SELECT COALESCE(username, ''), COALESCE(detector, ''),
+		         COALESCE(EXTRACT(EPOCH FROM recording_date)::BIGINT, 0)
+		         FROM uploads WHERE track_id = $1 LIMIT 1`
+	default:
+		query = `SELECT COALESCE(username, ''), COALESCE(detector, ''),
+		         COALESCE(recording_date, 0)
+		         FROM uploads WHERE track_id = ? LIMIT 1`
+	}
+	err := s.DB.DB.QueryRowContext(ctx, query, trackID).Scan(&username, &detector, &recordingDate)
+	if err != nil {
+		w.Header().Set("Content-Type", "application/json")
+		w.Header().Set("Cache-Control", "public, max-age=3600")
+		w.Write([]byte(`{"trackID":"` + trackID + `"}`))
+		return
+	}
+	w.Header().Set("Content-Type", "application/json")
+	w.Header().Set("Cache-Control", "public, max-age=3600")
+	json.NewEncoder(w).Encode(map[string]interface{}{
+		"trackID":       trackID,
+		"username":      username,
+		"detector":      detector,
+		"recordingDate": recordingDate,
+	})
+}
